feat: 댓글 비밀번호 bcrypt 기능 구현 1. 삭제 및 업데이트 로직 일부 수정

+++ blog_be/src/main/java/com/jumbochips/poml_jpa/comment/service/BlogCommentService.java
@@ -6,13 +6,13 @@ import com.jumbochips.poml_jpa.comment.domain.BlogComment;
 import com.jumbochips.poml_jpa.comment.dto.CommentRequestDto;
 import com.jumbochips.poml_jpa.comment.dto.CommentResponseDto;
 import com.jumbochips.poml_jpa.comment.repository.BlogCommentRepository;
-import com.jumbochips.poml_jpa.comment.repository.ProjectCommentRepository;
 import com.jumbochips.poml_jpa.common.recaptcha.RecaptchaValidationService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -20,6 +20,8 @@ public class BlogCommentService implements CommentService{
     private final BlogCommentRepository blogCommentRepository;
     private final BlogRepository blogRepository;
     private final RecaptchaValidationService recaptchaValidationService;
+    private final BCryptPasswordEncoder passwordEncoder;
+
 
     @Override
     public List<CommentResponseDto> getAllComment(Long blogId) {
@@ -43,10 +45,12 @@ public class BlogCommentService implements CommentService{
         Blog blog = blogRepository.findById(blogId)
                 .orElseThrow(() -> new IllegalArgumentException("Blog not found"));
 
+        String encodedPwd = passwordEncoder.encode(dto.getPassword());
+
         BlogComment comment = BlogComment.builder()
                 .blog(blog)
                 .username(dto.getUsername())
-                .pwd(dto.getPassword())
+                .pwd(encodedPwd)
                 .content(dto.getContent())
                 .build();
 
@@ -65,6 +69,10 @@ public class BlogCommentService implements CommentService{
         BlogComment comment = blogCommentRepository.findById(commentId)
                 .orElseThrow(() -> new IllegalArgumentException("Comment not found"));
 
+        if (!passwordEncoder.matches(dto.getPassword(), comment.getPwd())) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
+
         comment.updateContent(dto.getContent());
         blogCommentRepository.save(comment);
 
@@ -77,7 +85,13 @@ public class BlogCommentService implements CommentService{
     }
 
     @Override
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, String password) {
+        BlogComment comment = blogCommentRepository.findById(commentId)
+                        .orElseThrow(() -> new IllegalArgumentException("Comment not found"));
+
+        if (!passwordEncoder.matches(password, comment.getPwd())) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
         blogCommentRepository.deleteById(commentId);
     }
 }
